+++ core/src/lib.rs
@@ -22,8 +22,6 @@ pub mod alu;
 pub mod bytes;
 pub mod cpu;
 pub mod disassembler;
-#[deprecated(note = "Import from sp1_sdk instead of sp1_core")]
-pub mod io;
 pub mod lookup;
 pub mod memory;
 pub mod operations;
@@ -33,21 +31,6 @@ pub mod stark;
 pub mod syscall;
 pub mod utils;
 
-pub use io::*;
-
 #[allow(unused_imports)]
 use runtime::{Program, Runtime};
-use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-use stark::Proof;
 use stark::StarkGenericConfig;
-
-/// A proof of a RISCV ELF execution with given inputs and outputs.
-#[derive(Serialize, Deserialize)]
-#[deprecated(note = "Import from sp1_sdk instead of sp1_core")]
-pub struct SP1ProofWithIO<SC: StarkGenericConfig + Serialize + DeserializeOwned> {
-    #[serde(with = "proof_serde")]
-    pub proof: Proof<SC>,
-    pub stdin: SP1Stdin,
-    pub public_values: SP1PublicValues,
-}
